go/consensus/tendermint/abci: Move ApplicationState to its own file
